third_party/WebKit/LayoutTests/external/wpt/fullscreen/model/remove-single-manual-expected.txt

This is a testharness.js-based test.
FAIL Remove the single element on the fullscreen element stack assert_equals: expected Element node <div id="single"></div> but got Document node with 2 children
Harness: the test ran to completion.
